Move whole otherProfile page on to tableView.
